move stuff into constants file and utils file
