+++ src/sentry/tasks/integrations/github/open_pr_comment.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import itertools
 import logging
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from typing import Any, Dict, List, Set, Tuple
 
@@ -36,6 +35,17 @@ from sentry.silo.base import SiloMode
 from sentry.snuba.dataset import Dataset
 from sentry.snuba.referrer import Referrer
 from sentry.tasks.base import instrumented_task
+from sentry.tasks.integrations.github.constants import (
+    OPEN_PR_COMMENT_BODY_TEMPLATE,
+    OPEN_PR_ISSUE_DESCRIPTION_LENGTH,
+    OPEN_PR_ISSUE_ROW_TEMPLATE,
+    OPEN_PR_ISSUE_TABLE_TEMPLATE,
+    OPEN_PR_ISSUE_TABLE_TOGGLE_TEMPLATE,
+    OPEN_PR_MAX_FILES_CHANGED,
+    OPEN_PR_MAX_LINES_CHANGED,
+    OPEN_PR_METRICS_BASE,
+    STACKFRAME_COUNT,
+)
 from sentry.tasks.integrations.github.patch_parsers import PATCH_PARSERS
 from sentry.tasks.integrations.github.pr_comment import (
     ISSUE_LOCKED_ERROR_MESSAGE,
@@ -46,6 +56,7 @@ from sentry.tasks.integrations.github.pr_comment import (
     format_comment_url,
     get_pr_comment,
 )
+from sentry.tasks.integrations.github.utils import PullRequestFile
 from sentry.templatetags.sentry_helpers import small_count
 from sentry.types.referrer_ids import GITHUB_OPEN_PR_BOT_REFERRER
 from sentry.utils import metrics
@@ -53,57 +64,14 @@ from sentry.utils.snuba import raw_snql_query
 
 logger = logging.getLogger(__name__)
 
-OPEN_PR_METRICS_BASE = "github_open_pr_comment.{key}"
-
-
-@dataclass
-class PullRequestFile:
-    filename: str
-    patch: str
-
-
-# Caps the number of files that can be modified in a PR to leave a comment
-OPEN_PR_MAX_FILES_CHANGED = 7
-# Caps the number of lines that can be modified in a PR to leave a comment
-OPEN_PR_MAX_LINES_CHANGED = 500
-
-# Number of stackframes to check for filename + function combo, starting from the top
-STACKFRAME_COUNT = 6
-
-COMMENT_BODY_TEMPLATE = """## 🔍 Existing Issues For Review
-Your pull request is modifying functions with the following pre-existing issues:
-
-{issue_tables}
----
-
-<sub>Did you find this useful? React with a 👍 or 👎</sub>"""
-
-ISSUE_TABLE_TEMPLATE = """📄 File: **{filename}**
-
-| Function | Unhandled Issue |
-| :------- | :----- |
-{issue_rows}"""
-
-ISSUE_TABLE_TOGGLE_TEMPLATE = """<details>
-<summary><b>📄 File: {filename} (Click to Expand)</b></summary>
-
-| Function | Unhandled Issue |
-| :------- | :----- |
-{issue_rows}
-</details>"""
-
-ISSUE_ROW_TEMPLATE = "| **`{function_name}`** | [**{title}**]({url}) {subtitle} <br> `Event Count:` **{event_count}** |"
-
-ISSUE_DESCRIPTION_LENGTH = 52
-
 
 def format_open_pr_comment(issue_tables: List[str]) -> str:
-    return COMMENT_BODY_TEMPLATE.format(issue_tables="\n".join(issue_tables))
+    return OPEN_PR_COMMENT_BODY_TEMPLATE.format(issue_tables="\n".join(issue_tables))
 
 
 def format_open_pr_comment_subtitle(title_length, subtitle):
     # the title length + " " + subtitle should be <= 52
-    subtitle_length = ISSUE_DESCRIPTION_LENGTH - title_length - 1
+    subtitle_length = OPEN_PR_ISSUE_DESCRIPTION_LENGTH - title_length - 1
     return subtitle[: subtitle_length - 3] + "..." if len(subtitle) > subtitle_length else subtitle
 
 
@@ -111,7 +79,7 @@ def format_open_pr_comment_subtitle(title_length, subtitle):
 def format_issue_table(diff_filename: str, issues: List[PullRequestIssue], toggle=False) -> str:
     issue_rows = "\n".join(
         [
-            ISSUE_ROW_TEMPLATE.format(
+            OPEN_PR_ISSUE_ROW_TEMPLATE.format(
                 title=issue.title,
                 subtitle=format_open_pr_comment_subtitle(len(issue.title), issue.subtitle),
                 url=format_comment_url(issue.url, GITHUB_OPEN_PR_BOT_REFERRER),
@@ -123,9 +91,11 @@ def format_issue_table(diff_filename: str, issues: List[PullRequestIssue], toggl
     )
 
     if toggle:
-        return ISSUE_TABLE_TOGGLE_TEMPLATE.format(filename=diff_filename, issue_rows=issue_rows)
+        return OPEN_PR_ISSUE_TABLE_TOGGLE_TEMPLATE.format(
+            filename=diff_filename, issue_rows=issue_rows
+        )
 
-    return ISSUE_TABLE_TEMPLATE.format(filename=diff_filename, issue_rows=issue_rows)
+    return OPEN_PR_ISSUE_TABLE_TEMPLATE.format(filename=diff_filename, issue_rows=issue_rows)
 
 
 # for a single file, get the contents
