+++ src/sentry/tasks/integrations/github/pr_comment.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 
 import logging
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from enum import Enum
-from typing import Any, List, Optional
+from typing import Any, List
 
 import sentry_sdk
 from django.db import connection
@@ -13,7 +12,6 @@ from sentry_sdk.crons.decorator import monitor
 from snuba_sdk import Column, Condition, Direction, Entity, Function, Op, OrderBy, Query
 from snuba_sdk import Request as SnubaRequest
 
-from sentry.integrations.github.client import GitHubAppsClient
 from sentry.models.group import Group
 from sentry.models.groupowner import GroupOwnerType
 from sentry.models.options.organization_option import OrganizationOption
@@ -28,6 +26,14 @@ from sentry.snuba.dataset import Dataset
 from sentry.snuba.referrer import Referrer
 from sentry.tasks.base import instrumented_task
 from sentry.tasks.commit_context import DEBOUNCE_PR_COMMENT_CACHE_KEY
+from sentry.tasks.integrations.github.constants import (
+    ISSUE_LOCKED_ERROR_MESSAGE,
+    MERGED_PR_COMMENT_BODY_TEMPLATE,
+    MERGED_PR_METRICS_BASE,
+    MERGED_PR_SINGLE_ISSUE_TEMPLATE,
+    RATE_LIMITED_MESSAGE,
+)
+from sentry.tasks.integrations.github.utils import PullRequestIssue, create_or_update_comment
 from sentry.types.referrer_ids import GITHUB_PR_BOT_REFERRER
 from sentry.utils import metrics
 from sentry.utils.cache import cache
@@ -36,37 +42,11 @@ from sentry.utils.snuba import raw_snql_query
 
 logger = logging.getLogger(__name__)
 
-MERGED_PR_METRICS_BASE = "github_pr_comment.{key}"
-
-
-@dataclass
-class PullRequestIssue:
-    title: str
-    subtitle: str
-    url: str
-    affected_users: Optional[int] = None
-    event_count: Optional[int] = None
-    function_name: Optional[str] = None
-
 
 class GithubAPIErrorType(Enum):
     RATE_LIMITED = "gh_rate_limited"
     MISSING_PULL_REQUEST = "missing_gh_pull_request"
     UNKNOWN = "unknown_api_error"
-
-
-COMMENT_BODY_TEMPLATE = """## Suspect Issues
-This pull request was deployed and Sentry observed the following issues:
-
-{issue_list}
-
-<sub>Did you find this useful? React with a 👍 or 👎</sub>"""
-
-SINGLE_ISSUE_TEMPLATE = "- ‼️ **{title}** `{subtitle}` [View Issue]({url})"
-
-ISSUE_LOCKED_ERROR_MESSAGE = "Unable to create comment because issue is locked."
-
-RATE_LIMITED_MESSAGE = "API rate limit exceeded"
 
 
 def format_comment_subtitle(subtitle):
@@ -80,7 +60,7 @@ def format_comment_url(url, referrer):
 def format_comment(issues: List[PullRequestIssue]):
     issue_list = "\n".join(
         [
-            SINGLE_ISSUE_TEMPLATE.format(
+            MERGED_PR_SINGLE_ISSUE_TEMPLATE.format(
                 title=issue.title,
                 subtitle=format_comment_subtitle(issue.subtitle),
                 url=format_comment_url(issue.url, GITHUB_PR_BOT_REFERRER),
@@ -89,7 +69,7 @@ def format_comment(issues: List[PullRequestIssue]):
         ]
     )
 
-    return COMMENT_BODY_TEMPLATE.format(issue_list=issue_list)
+    return MERGED_PR_COMMENT_BODY_TEMPLATE.format(issue_list=issue_list)
 
 
 def pr_to_issue_query(pr_id: int):
@@ -153,49 +133,6 @@ def get_pr_comment(pr_id: int, comment_type: int) -> PullRequestComment | None:
         pull_request__id=pr_id, comment_type=comment_type
     )
     return pr_comment_query[0] if pr_comment_query.exists() else None
-
-
-def create_or_update_comment(
-    pr_comment: PullRequestComment | None,
-    client: GitHubAppsClient,
-    repo: Repository,
-    pr_key: int,
-    comment_body: str,
-    pullrequest_id: int,
-    issue_list: List[int],
-    comment_type: int = CommentType.MERGED_PR,
-    metrics_base=MERGED_PR_METRICS_BASE,
-):
-    # client will raise ApiError if the request is not successful
-    if pr_comment is None:
-        resp = client.create_comment(repo=repo.name, issue_id=pr_key, data={"body": comment_body})
-
-        current_time = timezone.now()
-        PullRequestComment.objects.create(
-            external_id=resp.body["id"],
-            pull_request_id=pullrequest_id,
-            created_at=current_time,
-            updated_at=current_time,
-            group_ids=issue_list,
-            comment_type=comment_type,
-        )
-        metrics.incr(metrics_base.format(key="comment_created"))
-    else:
-        resp = client.update_comment(
-            repo=repo.name, comment_id=pr_comment.external_id, data={"body": comment_body}
-        )
-        metrics.incr(metrics_base.format(key="comment_updated"))
-        pr_comment.updated_at = timezone.now()
-        pr_comment.group_ids = issue_list
-        pr_comment.save()
-
-    # TODO(cathy): Figure out a way to track average rate limit left for GH client
-
-    logger_event = metrics_base.format(key="create_or_update_comment")
-    logger.info(
-        logger_event,
-        extra={"new_comment": pr_comment is None, "pr_key": pr_key, "repo": repo.name},
-    )
 
 
 @instrumented_task(
